Pizza Api has Integrated[Listing,Create,Delete,Single Record]

+++ pizza-ordering-backend/app/Http/Controllers/PizzaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Pizza;
 use Illuminate\Http\Request;
+use App\Http\Resources\PizzasResource as PizzasCollection;
 use Illuminate\Support\Facades\Storage;
 
 class PizzaController extends Controller
@@ -15,9 +16,8 @@ class PizzaController extends Controller
      */
     public function index()
     {
-        $pizzas = Pizza::orderBy('id', 'desc')->get();
         $response = [
-            'pizzas_list' => $pizzas->toArray()
+            'pizzas_list' => PizzasCollection::collection(Pizza::orderBy('id', 'DESC')->get()),
         ];
         return response($response, 200);
     }
@@ -67,7 +67,7 @@ class PizzaController extends Controller
         $pizza = Pizza::find($id);
         if (!is_null($pizza)) {
             $response = [
-                'pizza' => $pizza->toArray()
+                'pizza' => new PizzasCollection($pizza)
             ];
             return response($response, 200);
         } else {
@@ -88,7 +88,7 @@ class PizzaController extends Controller
     public function update(Request $request, $id)
     {
         $pizza = Pizza::find($id);
-        $filePath='';//For holding the path of image_url
+        $filePath = ''; //For holding the path of image_url
         if (!is_null($pizza)) {
             $this->validate($request, [
                 'title' => 'required|string',
@@ -98,8 +98,7 @@ class PizzaController extends Controller
                 'price' => 'required|numeric',
                 'tax' => 'required|numeric'
             ]);
-            if($request->has('photo_url'))
-            {
+            if ($request->has('photo_url')) {
                 $this->validate($request, [
                     'photo_url' => 'required|mimes:png,jpg,jpeg|max:2048',
                 ]);
@@ -107,15 +106,11 @@ class PizzaController extends Controller
                 $filePath = $request->file('photo_url')->storeAs('uploads/pizza-images', $name, 'public');
                 $filePath = '/storage/' . $filePath;
                 Storage::delete($pizza->photo_url);
-            }
-            if($filePath)
-            {
                 $pizza['photo_url'] = $filePath;
             }
             $pizza['title'] = $request->title;
             $pizza['pizza_category_id'] = $request->pizza_category_id;
             $pizza['description'] = $request->description;
-            $pizza['photo_url'] =  $filePath;
             $pizza['ingredients'] = json_encode($request->ingredients);
             $pizza['tax'] = $request->tax;
             $pizza['price'] = $request->price;
@@ -142,11 +137,19 @@ class PizzaController extends Controller
     {
         $pizza = Pizza::find($id);
         if (!is_null($pizza)) {
-            $pizza->delete();
-            $response = [
-                'message' => 'The Pizza record has been deleted successfully'
-            ];
-            return response($response, 200);
+            if (is_null($pizza->orderDetails)) {
+                $pizza->delete();
+                $response = [
+                    'message' => 'The Pizza record been Deleted'
+                ];
+                return response($response, 200);
+            } else {
+                $response = [
+                    'message' => 'The Pizza record can not be Deleted',
+                    'status' => '409'
+                ];
+                return response($response, 200);
+            }
         } else {
             $response = [
                 'message' => 'The Pizza record has not been Founded'
